requirements.txt
toml
elasticsearch
python-dateutil
selenium